Create SocketNetmiko
Creates 1 socket to the bastion host (over ssh),
Over this single socket to the bastionserver, run netmiko library to a list of devices in multiple threads.

The reason for this script is that in the Netmiko(at least, I couldn't get it working efficiently)  jumphost example, it creates an ssh session per thread to the jumphost and then to the destination device. This makes the connections more efficient over the jumphost.